Added Tests and Comments as requested by CodeRabbitAI
